Split PlaybackService into section files
Turn playback/service.rs into a playback/service/ directory, relocating the
impl PlaybackService method clusters into sibling section modules (state,
pipeline, seek, advance, queue_commands, preview) and the unit tests into
tests.rs. mod.rs keeps the module doc, types, free helpers, constants, the
constructor, and the command dispatch loop.

Pure relocation: method bodies move verbatim. Moved methods gain a pub(super)
visibility prefix so the dispatch loop and sibling sections can still reach
them across module boundaries.
